feat: Decompose useDialFileManager into focused sub-hooks
