+++ openspec/changes/split-use-dial-file-manager/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-07-20
